moved sse4.1 to the right section and enabled sse4.1 tests which were for some reason not running because they weren't included in the module
