+++ src/x86_x64/sse4_1.rs
@@ -815,6 +815,28 @@ pub fn round_m128_s<const MODE: i32>(a: m128, b: m128) -> m128 {
   m128(unsafe { _mm_round_ss(a.0, b.0, MODE) })
 }
 
+/// Computes the bitwise AND of 256 bits in `a` and
+/// `b`, returns 1 if the result is zero, otherwise 0.
+/// * **Intrinsic:** [`_mm_testz_si128`]
+/// * **Assembly:**
+#[must_use]
+#[inline(always)]
+#[cfg_attr(docs_rs, doc(cfg(target_feature = "sse4.1")))]
+pub fn testz_m128i(a: m128i, b: m128i) -> i32 {
+  unsafe { _mm_testz_si128(a.0, b.0) }
+}
+
+/// Compute the bitwise NOT of `a` and then AND with `b`,
+/// returns 1 if the result is zero, otherwise 0.
+/// * **Intrinsic:** [`_mm_testc_si128`]
+/// * **Assembly:**
+#[must_use]
+#[inline(always)]
+#[cfg_attr(docs_rs, doc(cfg(target_feature = "sse4.1")))]
+pub fn testc_m128i(a: m128i, b: m128i) -> i32 {
+  unsafe { _mm_testc_si128(a.0, b.0) }
+}
+
 /// Tests if all bits are 1.
 ///
 /// ```
